Module P11_gcm_params

Parameter for AES GCM. In PKCS11 applications:

the length of iv can be any number between 1 and 256.

tag_bits can be any value between 0 and 128.

These invariants are not checked, in order to represent invalid states that DLLs may return.

type t

val equal : t -> t -> Ppx_deriving_runtime.bool

val compare : t -> t -> Ppx_deriving_runtime.int

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val make : iv:string -> aad:string -> tag_bits:P11_ulong.t -> t

val iv : t -> string

val aad : t -> string

val tag_bits : t -> P11_ulong.t